Additional wasm internal refactor
