+++ sdk/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/wasm/host/PureWasmHostFunctions.scala
@@ -5,7 +5,7 @@ package com.digitalasset.daml.lf.speedy.wasm
 
 package host
 
-import com.digitalasset.daml.lf.speedy.wasm.WasmUtils
+import com.digitalasset.daml.lf.speedy.wasm.internal.WasmUtils
 import com.dylibso.chicory.runtime.{HostFunction => WasmHostFunction}
 
 object PureWasmHostFunctions {
