+++ sdk/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/wasm/host/internal/WasmRunnerHostFunctions.scala
@@ -5,7 +5,7 @@ package com.digitalasset.daml.lf.speedy.wasm
 
 package host.internal
 
-import com.digitalasset.daml.lf.speedy.wasm.WasmUtils
+import com.digitalasset.daml.lf.speedy.wasm.internal.WasmUtils
 import com.dylibso.chicory.runtime.{HostFunction => WasmHostFunction, Instance => WasmInstance}
 import com.dylibso.chicory.wasm.types.{Value => WasmValue, ValueType => WasmValueType}
 import com.google.protobuf.ByteString
